backend/requirements.txt
Flask==3.1.2
pymongo==4.14.1
python-dotenv==1.1.1
Flask-Cors==6.0.1
gunicorn==23.0.0
requests==2.31.0
